refactor(executor): extract set operations into dedicated module (#4866)
Move UNION, INTERSECT, EXCEPT handling from execute.rs (1884 lines)
to a new set_operations.rs module (~590 lines), reducing execute.rs
to ~1307 lines.

Extracted components:
- execute_set_operations: Chain execution for set operations
- sort_set_operation_results: ORDER BY on set operation results
- resolve_order_by_column: Column resolution for set op ORDER BY
- extract_collations_from_select_list_with_schema: Collation helpers
- collect_union_aliases: Alias collection across UNION branches
- Helper functions for column name extraction

Closes #4866
